+++ src/components/Axis/AxisY.tsx
@@ -3,7 +3,13 @@ import React from 'react';
 import {axisLeft, axisRight, line, select} from 'd3';
 import type {Axis, AxisDomain, AxisScale, BaseType, Selection} from 'd3';
 
-import type {ChartScale, PreparedAxis, PreparedAxisPlotLine, PreparedSplit} from '../../hooks';
+import type {
+    ChartScale,
+    PreparedAxis,
+    PreparedAxisPlotBand,
+    PreparedAxisPlotLine,
+    PreparedSplit,
+} from '../../hooks';
 import {
     block,
     calculateCos,
@@ -11,6 +17,7 @@ import {
     formatAxisTickLabel,
     getAxisHeight,
     getAxisTitleRows,
+    getBandsPosition,
     getClosestPointsRange,
     getLineDashArray,
     getScaleTicks,
@@ -32,11 +39,15 @@ type Props = {
     height: number;
     split: PreparedSplit;
     plotRef?: React.MutableRefObject<SVGGElement | null>;
+    bottomLimit?: number;
 };
 
-function transformLabel(args: {node: Element; axis: PreparedAxis}) {
-    const {node, axis} = args;
+function transformLabel(args: {node: Element; axis: PreparedAxis; isTopOffsetOverload?: boolean}) {
+    const {node, axis, isTopOffsetOverload = false} = args;
     let topOffset = axis.labels.lineHeight / 2;
+    if (isTopOffsetOverload) {
+        topOffset = 0;
+    }
     let leftOffset = axis.labels.margin;
 
     if (axis.position === 'left') {
@@ -140,8 +151,20 @@ type PlotLineData = {
     transform: string;
 } & PreparedAxisPlotLine;
 
+type PlotBandData = {
+    transform: string;
+} & PreparedAxisPlotBand;
+
 export const AxisY = (props: Props) => {
-    const {axes, width, height: totalHeight, scale, split, plotRef} = props;
+    const {
+        axes: allAxes,
+        width,
+        height: totalHeight,
+        scale,
+        split,
+        plotRef,
+        bottomLimit = 0,
+    } = props;
     const height = getAxisHeight({split, boundsHeight: totalHeight});
     const ref = React.useRef<SVGGElement | null>(null);
     const lineGenerator = line();
@@ -151,8 +174,18 @@ export const AxisY = (props: Props) => {
             return;
         }
 
+        const axes = allAxes.filter((a) => a.visible);
+
         const svgElement = select(ref.current);
         svgElement.selectAll('*').remove();
+
+        let plotContainer = null;
+        const plotClassName = b('plot-y');
+
+        if (plotRef?.current) {
+            plotContainer = select(plotRef.current);
+            plotContainer.selectAll(`.${plotClassName}`).remove();
+        }
 
         const getAxisPosition = (axis: PreparedAxis) => {
             const top = split.plots[axis.plotIndex]?.top || 0;
@@ -172,6 +205,19 @@ export const AxisY = (props: Props) => {
                             transform: getAxisPosition(axis),
                         };
                     }),
+                );
+            }
+
+            return acc;
+        }, []);
+
+        const plotBands = axes.reduce<PlotBandData[]>((acc, axis) => {
+            if (axis.plotBands.length) {
+                acc.push(
+                    ...axis.plotBands.map((plotBand) => ({
+                        ...plotBand,
+                        transform: getAxisPosition(axis),
+                    })),
                 );
             }
 
@@ -204,8 +250,8 @@ export const AxisY = (props: Props) => {
             yAxisGenerator(axisItem);
 
             if (d.labels.enabled) {
-                const tickTexts = axisItem
-                    .selectAll<SVGTextElement, string>('.tick text')
+                const labels = axisItem.selectAll<SVGTextElement, string>('.tick text');
+                const tickTexts = labels
                     // The offset must be applied before the labels are rotated.
                     // Therefore, we reset the values and make an offset in transform  attribute.
                     // FIXME: give up axisLeft(d3) and switch to our own generation method
@@ -215,6 +261,29 @@ export const AxisY = (props: Props) => {
                     .style('transform', function () {
                         return transformLabel({node: this, axis: d});
                     });
+
+                labels.each(function (_d, i) {
+                    if (i === 0) {
+                        const currentElement = this as SVGTextElement;
+                        const currentElementPosition = currentElement.getBoundingClientRect();
+                        const text = select(currentElement);
+
+                        if (currentElementPosition.bottom > bottomLimit) {
+                            const transform = transformLabel({
+                                node: this,
+                                axis: d,
+                                isTopOffsetOverload: true,
+                            });
+                            text.style('transform', transform);
+                            if (d.labels.rotation) {
+                                text.attr('text-anchor', () => {
+                                    return d.labels.rotation < 0 ? 'start' : 'end';
+                                });
+                            }
+                        }
+                    }
+                });
+
                 const textMaxWidth =
                     !d.labels.rotation || Math.abs(d.labels.rotation) % 360 !== 90
                         ? d.labels.maxWidth
@@ -241,16 +310,56 @@ export const AxisY = (props: Props) => {
                     .remove();
             }
 
-            if (plotRef && d.plotLines.length > 0) {
-                const plotLineClassName = b('plotLine');
-                const plotLineContainer = select(plotRef.current);
-                plotLineContainer.selectAll(`.${plotLineClassName}`).remove();
+            if (plotContainer && d.plotBands.length > 0) {
+                const plotBandClassName = b('plot-y-band');
 
-                const plotLinesSelection = plotLineContainer
+                const plotBandsSelection = plotContainer
+                    .selectAll(`.${plotBandClassName}`)
+                    .data(plotBands)
+                    .join('g')
+                    .attr('class', `${plotClassName} ${plotBandClassName}`)
+                    .style('transform', (plotBand) => plotBand.transform);
+
+                plotBandsSelection
+                    .append('rect')
+                    .attr('x', 0)
+                    .attr('width', width)
+                    .attr('y', (band) => {
+                        const {from, to} = getBandsPosition({band, axisScale, axis: 'y'});
+                        const halfBandwidth = (axisScale.bandwidth?.() ?? 0) / 2;
+                        const startPos = halfBandwidth + Math.min(from, to);
+
+                        return Math.max(0, startPos);
+                    })
+                    .attr('height', (band) => {
+                        const {from, to} = getBandsPosition({band, axisScale, axis: 'y'});
+                        const startPos = height - Math.min(from, to);
+                        const endPos = Math.min(Math.abs(to - from), startPos);
+
+                        return Math.min(endPos, height);
+                    })
+                    .attr('fill', (band) => band.color)
+                    .attr('opacity', (band) => band.opacity);
+
+                plotBandsSelection.each((plotBandData, i, nodes) => {
+                    const plotLineSelection = select(nodes[i]);
+
+                    if (plotBandData.layerPlacement === 'before') {
+                        plotLineSelection.lower();
+                    } else {
+                        plotLineSelection.raise();
+                    }
+                });
+            }
+
+            if (plotContainer && d.plotLines.length > 0) {
+                const plotLineClassName = b('plot-y-line');
+
+                const plotLinesSelection = plotContainer
                     .selectAll(`.${plotLineClassName}`)
                     .data(plotLines)
                     .join('g')
-                    .attr('class', plotLineClassName)
+                    .attr('class', `${plotClassName} ${plotLineClassName}`)
                     .style('transform', (plotLine) => plotLine.transform);
 
                 plotLinesSelection
@@ -332,7 +441,7 @@ export const AxisY = (props: Props) => {
                     handleOverflowingText(nodes[index] as SVGTSpanElement, height);
                 }
             });
-    }, [axes, width, height, scale, split]);
+    }, [allAxes, width, height, scale, split, bottomLimit]);
 
     return <g ref={ref} className={b('container')} />;
 };