+++ src/hooks/useChartDimensions/utils.ts
@@ -16,6 +16,10 @@ export const getBoundsWidth = (args: {
 };
 
 export function getYAxisWidth(axis: PreparedAxis | undefined) {
+    if (!axis?.visible) {
+        return 0;
+    }
+
     let result = 0;
     if (axis?.title.text) {
         result += axis.title.height + axis.title.margin;